Revert "Merge branch 'main' into develop-carlos"
This reverts commit 59b61e2e2bcdf4b17a69077b2d1cdac736613cc5, reversing
changes made to b505493c18ba4f7f0220ed4635a39b6256b30c5b.

+++ lib/features/user/data/datasources/user_remote_data_source.dart
@@ -23,7 +23,7 @@ abstract class UserRemoteDataSource {
 
 class UserRemoteDataSourceImpl implements UserRemoteDataSource {
   final Dio dioClient = Dio();
-  final String apiUrl = 'http://192.168.100.40:3000'; // URL de tu API
+  final String apiUrl = 'http://localhost:3000'; // URL de tu API
   final Session session = Session.instance;
 
   UserRemoteDataSourceImpl();
@@ -48,10 +48,9 @@ class UserRemoteDataSourceImpl implements UserRemoteDataSource {
           name: response.data['nombre'],
           email: response.data['correo'],
           password: response.data['contrasenia'],
-          userType: response.data['tipoUsuario'],
-            photo: response.data['foto'] != null
-              ? Uint8List.fromList(List<int>.from(response.data['foto']['data']))
-              : Uint8List(0), // Si no hay foto, crea un Uint8List vacío
+          userType: response.data['tipo'],
+          photo: Uint8List.fromList(List<int>.from(response.data['foto'][
+              'data'])), // Convierte la foto a Uint8List si es un arreglo de bytes
           disponibility: response.data['disponibilidad'] == 1,
           location: response.data['ubicacion'] ?? '',
         );
@@ -59,11 +58,9 @@ class UserRemoteDataSourceImpl implements UserRemoteDataSource {
             userId: idUsuario.toString(), token: token, user: userAuth);
         return userAuth;
       } catch (e) {
-        print('Failed to authenticate user1');
-        throw Exception('Failed to authenticate user1');        
+        throw Exception('Failed to authenticate user1');
       }
     } else {
-      print('Failed to authenticate user2');
       throw Exception('Failed to authenticate user2');
     }
   }
@@ -73,7 +70,7 @@ class UserRemoteDataSourceImpl implements UserRemoteDataSource {
 
     try {
       final response = await dioClient.post(
-        '$apiUrl/users',
+        '$apiUrl/usuarios',
         data: userModel.toFormData(),
         options: Options(
           validateStatus: (status) => status! < 500,
@@ -110,10 +107,14 @@ class UserRemoteDataSourceImpl implements UserRemoteDataSource {
       'foto': MultipartFile.fromBytes(photo, filename: 'photo.jpg'),
     });
 
+    String? userId = session.userId;
+    if (userId == null) {
+      throw Exception('User ID is not available');
+    }
 
     try {
       final response = await dioClient.put(
-        '$apiUrl/users',
+        '$apiUrl/usuarios/$userId',
         data: formData,
         options: Options(
           validateStatus: (status) {
@@ -153,10 +154,13 @@ class UserRemoteDataSourceImpl implements UserRemoteDataSource {
 
   @override
   Future<bool> deleteUser() async {
-    
+    String? userId = session.userId;
+    if (userId == null) {
+      throw Exception('User ID is not available');
+    }
     try {
       final response = await dioClient.delete(
-        '$apiUrl/users',
+        '$apiUrl/usuarios',
         options: Options(
           headers: {
             'Content-Type': 'multipart/form-data',
